+++ jest.config.js
@@ -20,7 +20,7 @@ module.exports = {
       branches: 100,
       functions: 100,
       lines: 100,
-      statements: 0
+      statements: 100
     }
   }
 }